log package refactor
