+++ src/main/java/com/configcat/ConfigCatHooks.java
@@ -62,7 +62,7 @@ public class ConfigCatHooks {
         }
     }
 
-    void invokeOnClientReady() {
+    public void invokeOnClientReady() {
         synchronized (sync) {
             for (Runnable func : this.onClientReady) {
                 func.run();
@@ -70,7 +70,7 @@ public class ConfigCatHooks {
         }
     }
 
-    void invokeOnError(String error) {
+    public void invokeOnError(String error) {
         synchronized (sync) {
             for (Consumer<String> func : this.onError) {
                 func.accept(error);
@@ -78,7 +78,7 @@ public class ConfigCatHooks {
         }
     }
 
-    void invokeOnConfigChanged(Map<String, Setting> settingMap) {
+    public void invokeOnConfigChanged(Map<String, Setting> settingMap) {
         synchronized (sync) {
             for (Consumer<Map<String, Setting>> func : this.onConfigChanged) {
                 func.accept(settingMap);
@@ -86,7 +86,7 @@ public class ConfigCatHooks {
         }
     }
 
-    void invokeOnFlagEvaluated(EvaluationDetails<Object> evaluationDetails) {
+    public void invokeOnFlagEvaluated(EvaluationDetails<Object> evaluationDetails) {
         synchronized (sync) {
             for (Consumer<EvaluationDetails<Object>> func : this.onFlagEvaluated) {
                 func.accept(evaluationDetails);
@@ -94,7 +94,7 @@ public class ConfigCatHooks {
         }
     }
 
-    void clear() {
+    public void clear() {
         synchronized (sync) {
             this.onConfigChanged.clear();
             this.onError.clear();
