+++ src/main/java/com/configcat/OverrideDataSourceBuilder.java
@@ -1,5 +1,7 @@
 package com.configcat;
 
+import com.configcat.log.ConfigCatLogger;
+
 import java.util.Map;
 
 /**
