+++ src/test/java/com/configcat/DataGovernanceTest.java
@@ -1,5 +1,6 @@
 package com.configcat;
 
+import com.configcat.log.ConfigCatLogger;
 import okhttp3.OkHttpClient;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
